requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-generativeai